Bug Fix: lastMoveMade needed to be cast to a Number
It was set using Array .slice(-1) so it was an array reference and was not working when I tried using it to look up affected lines, as the Map.get method of the cellToLinesMap expects a number, not an array reference.  I wonder if using Array.pop() instead of slice would have avoided this problem.