Make modal generate question. App knows only if correct.
Formerly the App component generated a question and passed it to the questionModal.  Instead of all that info flowing down we now have the modal generate it's question, manage input state, and check submissions internally.  The only thing it gets from the app is a handleAnswerSubmit method that enables the modal to inform the app if the submission was correct or not.  The App closes the modal and handles state that has to do with the "turn" but not the "question" such as the active cell.