Modify questionGenerator for different params
Now takes mathTopics and questionsRightSoFar and uses these to pick own topic and difficulty level instead of making the App do that and pass the topic and difficulty to the generator.  this groups logic more sensibly. Picking a difficulty based on a score is part of generating a question, not core functionality of the overarching App.