Solve double submit issue
There was an error occurring when a player BOTH clicked the submit button and hit the Enter key to submit.  The first call to handleAnswerSubmit would reset the players answer and the second call would change the banner from "correct" to "No".  I tried fixing this by adding a boolean flag to the state to track whether or not submit had been clicked but this did not solve the issue.  In the end, I solved the issue in a less than ideal way, by totally disabling the Enter button as a way to submit answers.