Reenable enter key to submit
Problem of Double submit exists with or without the enter key functionality.  Double clicking the Submit Button causes the same error that clicking it and hitting "enter" does.  This needs to be fixed by adding something to the handle submit method in App.js that returns early if submit is clicked twice withing the two seconds that it takes to display a modal close and chip drop.